requirements.txt
pandas==2.1.4
boto3==1.34.0
opensearch-py==2.4.2
pydantic==2.5.3
python-dotenv==1.0.0
schedule==1.2.0
kagglehub==0.2.5